requirements.txt
robotframework==7.0
robotframework-browser==18